Move new export hooks into tool-specific subsystem files.
[ci skip-rust]

[ci skip-build-wheels]
