+++ src/python/pants/backend/python/lint/docformatter/subsystem.py
@@ -3,6 +3,7 @@
 
 
 from pants.backend.python.goals import lockfile
+from pants.backend.python.goals.export import ExportPythonTool, ExportPythonToolSentinel
 from pants.backend.python.goals.lockfile import GeneratePythonLockfile
 from pants.backend.python.subsystems.python_tool_base import PythonToolBase
 from pants.backend.python.subsystems.setup import PythonSetup
@@ -47,9 +48,21 @@ def setup_lockfile_request(
     )
 
 
+class DocformatterExportSentinel(ExportPythonToolSentinel):
+    pass
+
+
+@rule
+def docformatter_export(
+    _: DocformatterExportSentinel, docformatter: Docformatter
+) -> ExportPythonTool:
+    return ExportPythonTool(tool_name=docformatter.name, pex_request=docformatter.to_pex_request())
+
+
 def rules():
     return (
         *collect_rules(),
         *lockfile.rules(),
         UnionRule(GenerateToolLockfileSentinel, DocformatterLockfileSentinel),
+        UnionRule(ExportPythonToolSentinel, DocformatterExportSentinel),
     )
